test_bfs_new is nx-free refactoring, 2 tests are skipped
